+++ src/components/ui/action/Action.styled.ts
@@ -1,13 +1,13 @@
 import type { ThemeCSS } from "../../../theme";
 
-import { fontPxToRem, borderPxToRem, dimensionsPxToRem, styled } from "../../../theme";
+import { borderPxToRem, dimensionsPxToRem, fontPxToRem, styled } from "../../../theme";
 
 const StyledAction = styled("div", {
     height: dimensionsPxToRem(102),
     width: dimensionsPxToRem(102),
     borderRadius: dimensionsPxToRem(102),
     border: `${borderPxToRem(1)} solid $buttonBorder`,
-    background: "white",
+    background: "$background",
     padding: 0,
     display: "flex",
     alignItems: "center",
@@ -18,7 +18,7 @@ const StyledAction = styled("div", {
 const Badge = styled("div", {
     position: "absolute",
     background: "$red1",
-    color: "white",
+    color: "$background",
     fontWeight: "bold",
     top: "-25%",
     right: "-25%",
@@ -28,9 +28,9 @@ const Badge = styled("div", {
     padding: dimensionsPxToRem(17),
     width: dimensionsPxToRem(60),
     height: dimensionsPxToRem(60),
-    alignItems: "center", // Vertically center content
-    justifyContent: "center", // Horizontally center content
-    borderRadius: "50%", // Make it circular
+    alignItems: "center",
+    justifyContent: "center",
+    borderRadius: "50%",
 });
 
 const sharedCss: ThemeCSS = {